:card_file_box: DB related changes
