+++ backend/router/api/auth.py
@@ -1,7 +1,8 @@
 """
 url route split
 """
-from flask import Blueprint, request, jsonify, current_app
+import traceback
+from flask import Blueprint, request, jsonify, session, current_app
 from util.logger import logger
 from model import user_model
 
@@ -11,6 +12,76 @@ import datetime
 auth_route = Blueprint('auth_route', __name__)
 
 
+"""
+@app.route("/api/auth/signup", methods=['POST'])
+def auth_signup():
+    logger.info("SignUp!")
+    logger.info(request.get_json())
+
+    data = request.get_json()
+
+    username = data.get('username')
+    useremail = data.get('useremail')
+    userpwd = data.get('userpwd')
+    bio = data.get('bio')
+
+    user_data = user_model.User.query.filter_by(username=username).first()
+    if user_data is not None:
+        logger.info("Username is Already exist")
+        return {"success": "username is already exist"}
+
+    try:
+        user = user_model.User(**data)
+        user.has_password()
+        user_model.db.session.add(user)
+        user_model.db.session.commit()
+        user_model.db.session.remove()
+        logger.info("user Save Success")
+        return {'success': 'user Save Success'}
+    except Exception as e:
+        logger.error("user Save Fail")
+        logger.debug(traceback.print_exc(e))
+        return {'fail': "user Save Fail"}
+
+
+@app.route("/api/auth/login", methods=['POST'])
+def auth_login():
+    logger.info("User auth Login")
+    logger.info(request.get_json())
+
+    username = request.get_json()['username']
+    useremail = request.get_json().get('useremail')
+    userpwd = request.get_json()['userpwd']
+
+    my_user = user_model.User()
+
+    try:
+        user_data = my_user.query.filter_by(username=username).first()
+
+        if user_data is not None:
+            auth = user_data.check_password(userpwd)
+            if not auth:
+                logger.info("password validation fail!")
+                return {'status': 'fail'}, 401
+            else:
+                logger.info("login success!")
+                session['login'] = True
+                return {'success': session['login']}, 200
+        else:
+            logger.info("user information is wrong or user does  not exists....")
+            return {'status': 'fail'}, 401
+    except Exception as e:
+        logger.error("login Exception...")
+        logger.debug(traceback.print_exc(e))
+        return {'status': 'fail'}, 404
+
+
+@app.route('/api/auth/logout')
+def auth_logout():
+    session['login'] = False
+    return {'success': 'logout'}
+
+"""
 @auth_route.route('/register', methods=['POST'])
 def register_user():
     logger.info("user register")
@@ -21,7 +92,6 @@ def register_user():
     if user_data is not None:
         logger.error("Username already exists")
         return {"failed": "username is already exist"}
-
     user = user_model.User(**data)
     user.has_password()
     db.session.add(user)
