+++ db/create_users.sql
@@ -1,10 +1,8 @@
 CREATE USER 'root'@'172.18.0.1' IDENTIFIED BY '1010';
-CREATE USER 'jahn'@'%' IDENTIFIED BY '1010';
-CREATE USER 'jahn'@'172.18.0.1' IDENTIFIED BY '1010';
+-- CREATE USER 'jahn'@'%' IDENTIFIED BY '1010';
+-- CREATE USER 'jahn'@'172.18.0.1' IDENTIFIED BY '1010';
 GRANT ALL PRIVILEGES ON *.* TO 'root'@'172.18.0.1';
-GRANT ALL PRIVILEGES ON *.* TO 'jahn'@'%';
-GRANT ALL PRIVILEGES ON *.* TO 'jahn'@'172.18.0.1';
+-- GRANT ALL PRIVILEGES ON *.* TO 'jahn'@'%';
+-- GRANT ALL PRIVILEGES ON *.* TO 'jahn'@'172.18.0.1';
 
 flush privileges;
-
--- Dump completed on 2021-01-06  9:37:51
